[TS-376] Refactor: footer btn refactoring
