+++ src/components/Habit/RestRecord/RestRecord.style.ts
@@ -8,7 +8,7 @@ export const layoutStyle = css`
 	gap: 40px;
 	width: 22.5rem;
 	margin: 0 auto;
-	padding: 4.75rem 1.5rem 0;
+	padding: 4.75rem 1.5rem 9.5rem;
 
 	& > h1 {
 		${theme.font.head_a}
